new/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==5.0.1
gunicorn==21.2.0
paho-mqtt==2.1.0